+++ src/Domain/DependencyInjection/index.ts
@@ -1,2 +1,3 @@
 export * from "./ServiceDescriptor";
 export * from "./Injectable";
+export * from "./Inject";